Even more flake8. Chagne Constraints into inheriting from Operator, which makes things a little cleaner and opens up more avenues for using eflags in algebra.
